requirements.txt
python-telegram-bot==20.3
pytz==2023.3
aiohttp==3.8.4